Split the tests into Recorder and DataFormat tests
